+++ lending/src/sdk/sdk.ts
@@ -3,11 +3,23 @@
  */
 
 import { ClientSDK } from "../lib/sdks.js";
-import { Pets } from "./pets.js";
+import { Pet } from "./pet.js";
+import { Store } from "./store.js";
+import { User } from "./user.js";
 
 export class LendingSDK extends ClientSDK {
-  private _pets?: Pets;
-  get pets(): Pets {
-    return (this._pets ??= new Pets(this._options));
+  private _pet?: Pet;
+  get pet(): Pet {
+    return (this._pet ??= new Pet(this._options));
+  }
+
+  private _store?: Store;
+  get store(): Store {
+    return (this._store ??= new Store(this._options));
+  }
+
+  private _user?: User;
+  get user(): User {
+    return (this._user ??= new User(this._options));
   }
 }